pyproject.toml
[project]
name = "walletapp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi (>=0.115.8,<0.116.0)",
    "pydantic (>=2.10.6,<3.0.0)",
    "sqlalchemy (>=2.0.38,<3.0.0)",
    "dependency-injector>=4.45.0,<5.0.0",
    "uvicorn (>=0.34.0,<0.35.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "alembic (>=1.14.1,<2.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "gunicorn (>=23.0.0,<24.0.0)",
    "pytest-asyncio (>=0.25.3,<0.26.0)",
    "ruff>=0.11.0",
]
